+++ includes/content.php
@@ -14,12 +14,6 @@ class Fiber_Admin_Content{
 			add_filter('the_content', array($this, 'fiad_auto_convert_email_address'));
 		}
 		
-		// Enable SVG
-		if(fiad_get_miscellaneous_option('enable_svg')){
-			add_filter('upload_mimes', array($this, 'fiad_svg_mime_types'));
-			add_action('admin_head', array($this, 'fiad_fix_svg_display'));
-		}
-		
 		// Content protection
 		if(!fiad_get_miscellaneous_option('disable_content_protection') && !fiad_is_admin_user_role()){
 			add_action('wp_footer', array($this, 'fiad_content_protection_scripts'));
@@ -59,21 +53,6 @@ class Fiber_Admin_Content{
 		$replace = array('<a href="mailto:$1">$1</a>');
 		
 		return preg_replace($search, $replace, $content);
-	}
-	
-	public function fiad_svg_mime_types($mimes){
-		$mimes['svg'] = 'image/svg+xml';
-		
-		return $mimes;
-	}
-	
-	public function fiad_fix_svg_display(){
-		echo '<style>
-		    td.media-icon img[src$=".svg"], img[src$=".svg"].attachment-post-thumbnail{
-		      width: 100% !important;
-		      height: auto !important;
-		    }
-		  </style>';
 	}
 	
 	public function fiad_content_protection_scripts(){